Fixes grid resizing issue with fixed height
The javascript had a fixed minimum value for the box height. This caused
problems with the aspect ratio of the grid container. Some of the flex
properties were also set on the grid row and grid container flex boxes
to ensure correct scaling beahviour.

The grid container is now inside a main container and there's a header
above it.